requirements.txt
# Telegram网盘机器人依赖
python-telegram-bot==20.3
apscheduler==3.10.4
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
uvicorn==0.25.0
fastapi==0.109.0
pydantic==2.5.3
httpx==0.26.0
Pillow==10.2.0
python-magic==0.4.27
